feat(tested): running example
Since the training is really complicated it was said by Nori that the researcher will not provide support for this one. Instead the serve needed to be tested. After some works we completed the test in our platform by changing part of the code since we had some problem because of the ctc_decode function. Changed it from the